Added feature to allow the user to specify a list of files via @LISTFILE. Documentation and code cleanup to follow.
Modified _writeline() to take a hash instead of a bunch of individual
params.

Corrected a bug where loaded manifests would not correctly interpret the
octal mode string from a file.  The value was only used internally, so the
bug didn't reveal itself until the @LISTFILE feature was added and the two
methods of getting a manifest hash (generated vs loaded) created conflicting
modes.